Moved two integrators into separate files
