+++ platforms/common/include/openmm/common/CommonKernels.h
@@ -33,8 +33,9 @@
 #include "openmm/Platform.h"
 #include "openmm/kernels.h"
 #include "openmm/internal/CompiledExpressionSet.h"
-#include "openmm/internal/CustomIntegratorUtilities.h"
+#include "openmm/internal/ContextImpl.h"
 #include "lepton/CompiledExpression.h"
+#include "lepton/CustomFunction.h"
 #include "lepton/ExpressionProgram.h"
 
 namespace OpenMM {
@@ -942,118 +943,6 @@ private:
     ComputeKernel kernel1, kernel2, kernel3;
 };
 
-/*
- * This kernel is invoked by NoseHooverIntegrator to take one time step.
- */
-class CommonIntegrateNoseHooverStepKernel : public IntegrateNoseHooverStepKernel {
-public:
-    CommonIntegrateNoseHooverStepKernel(std::string name, const Platform& platform, ComputeContext& cc) :
-                                  IntegrateNoseHooverStepKernel(name, platform), cc(cc), hasInitializedKernels(false),
-                                  hasInitializedKineticEnergyKernel(false), hasInitializedHeatBathEnergyKernel(false),
-                                  hasInitializedScaleVelocitiesKernel(false), hasInitializedPropagateKernel(false) {}
-    ~CommonIntegrateNoseHooverStepKernel() {}
-    /**
-     * Initialize the kernel.
-     *
-     * @param system     the System this kernel will be applied to
-     * @param integrator the NoseHooverIntegrator this kernel will be used for
-     */
-    void initialize(const System& system, const NoseHooverIntegrator& integrator);
-    /**
-     * Execute the kernel.
-     *
-     * @param context    the context in which to execute this kernel
-     * @param integrator the VerletIntegrator this kernel is being used for
-     */
-    void execute(ContextImpl& context, const NoseHooverIntegrator& integrator);
-    /**
-     * Compute the kinetic energy.
-     *
-     * @param context    the context in which to execute this kernel
-     * @param integrator the NoseHooverIntegrator this kernel is being used for
-     */
-    double computeKineticEnergy(ContextImpl& context, const NoseHooverIntegrator& integrator);
-    /**
-     * Execute the kernel that propagates the Nose Hoover chain and determines the velocity scale factor.
-     * 
-     * @param context  the context in which to execute this kernel
-     * @param noseHooverChain the object describing the chain to be propagated.
-     * @param kineticEnergy the {center of mass, relative} kineticEnergies of the particles being thermostated by this chain.
-     * @param timeStep the time step used by the integrator.
-     * @return the velocity scale factor to apply to the particles associated with this heat bath.
-     */
-    std::pair<double, double> propagateChain(ContextImpl& context, const NoseHooverChain &noseHooverChain, std::pair<double, double> kineticEnergy, double timeStep);
-    /**
-     * Execute the kernal that computes the total (kinetic + potential) heat bath energy.
-     *
-     * @param context the context in which to execute this kernel
-     * @param noseHooverChain the chain whose energy is to be determined.
-     * @return the total heat bath energy.
-     */
-    double computeHeatBathEnergy(ContextImpl& context, const NoseHooverChain &noseHooverChain);
-    /**
-     * Execute the kernel that computes the kinetic energy for a subset of atoms,
-     * or the relative kinetic energy of Drude particles with respect to their parent atoms
-     *
-     * @param context the context in which to execute this kernel
-     * @param noseHooverChain the chain whose energy is to be determined.
-     * @param downloadValue whether the computed value should be downloaded and returned.
-     */
-    std::pair<double, double> computeMaskedKineticEnergy(ContextImpl& context, const NoseHooverChain &noseHooverChain, bool downloadValue);
-    /**
-     * Execute the kernel that scales the velocities of particles associated with a nose hoover chain
-     *
-     * @param context the context in which to execute this kernel
-     * @param noseHooverChain the chain whose energy is to be determined.
-     * @param scaleFactor the multiplicative factor by which {absolute, relative} velocities are scaled.
-     */
-    void scaleVelocities(ContextImpl& context, const NoseHooverChain &noseHooverChain, std::pair<double, double> scaleFactor);
-    /**
-     * Write the chain states to a checkpoint.
-     */
-    void createCheckpoint(ContextImpl& context, std::ostream& stream) const;
-    /**
-     * Load the chain states from a checkpoint.
-     */
-    void loadCheckpoint(ContextImpl& context, std::istream& stream);
-    /**
-     * Get the internal states of all chains.
-     * 
-     * @param context       the context for which to get the states
-     * @param positions     element [i][j] contains the position of bead j for chain i
-     * @param velocities    element [i][j] contains the velocity of bead j for chain i
-     */
-    void getChainStates(ContextImpl& context, std::vector<std::vector<double> >& positions, std::vector<std::vector<double> >& velocities) const;
-    /**
-     * Set the internal states of all chains.
-     * 
-     * @param context       the context for which to get the states
-     * @param positions     element [i][j] contains the position of bead j for chain i
-     * @param velocities    element [i][j] contains the velocity of bead j for chain i
-     */
-    void setChainStates(ContextImpl& context, const std::vector<std::vector<double> >& positions, const std::vector<std::vector<double> >& velocities);
-private:
-    ComputeContext& cc;
-    float prevMaxPairDistance;
-    ComputeArray maxPairDistanceBuffer, pairListBuffer, atomListBuffer, pairTemperatureBuffer, oldDelta;
-    std::map<int, ComputeArray> chainState;
-    ComputeKernel kernel1, kernel2, kernel3, kernel4, kernelHardWall;
-    bool hasInitializedKernels;
-    ComputeKernel reduceEnergyKernel;
-    ComputeKernel computeHeatBathEnergyKernel;
-    ComputeKernel computeAtomsKineticEnergyKernel;
-    ComputeKernel computePairsKineticEnergyKernel;
-    ComputeKernel scaleAtomsVelocitiesKernel;
-    ComputeKernel scalePairsVelocitiesKernel;
-    ComputeArray energyBuffer, scaleFactorBuffer, kineticEnergyBuffer, chainMasses, chainForces, heatBathEnergy;
-    std::map<int, ComputeArray> atomlists, pairlists;
-    std::map<int, ComputeKernel> propagateKernels;
-    bool hasInitializedPropagateKernel;
-    bool hasInitializedKineticEnergyKernel;
-    bool hasInitializedHeatBathEnergyKernel;
-    bool hasInitializedScaleVelocitiesKernel;
-};
-
 /**
  * This kernel is invoked by BrownianIntegrator to take one time step.
  */
@@ -1166,132 +1055,6 @@ private:
     ComputeArray params, oldDelta;
     ComputeKernel kernel1, kernel2, kernel3, selectSizeKernel;
     double prevTemp, prevFriction, prevErrorTol;
-};
-
-/**
- * This kernel is invoked by CustomIntegrator to take one time step.
- */
-class CommonIntegrateCustomStepKernel : public IntegrateCustomStepKernel {
-public:
-    enum GlobalTargetType {DT, VARIABLE, PARAMETER};
-    CommonIntegrateCustomStepKernel(std::string name, const Platform& platform, ComputeContext& cc) : IntegrateCustomStepKernel(name, platform), cc(cc),
-            hasInitializedKernels(false), deviceGlobalsAreCurrent(false), needsEnergyParamDerivs(false) {
-    }
-    /**
-     * Initialize the kernel.
-     * 
-     * @param system     the System this kernel will be applied to
-     * @param integrator the CustomIntegrator this kernel will be used for
-     */
-    void initialize(const System& system, const CustomIntegrator& integrator);
-    /**
-     * Execute the kernel.
-     * 
-     * @param context    the context in which to execute this kernel
-     * @param integrator the CustomIntegrator this kernel is being used for
-     * @param forcesAreValid if the context has been modified since the last time step, this will be
-     *                       false to show that cached forces are invalid and must be recalculated.
-     *                       On exit, this should specify whether the cached forces are valid at the
-     *                       end of the step.
-     */
-    void execute(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid);
-    /**
-     * Compute the kinetic energy.
-     * 
-     * @param context    the context in which to execute this kernel
-     * @param integrator the CustomIntegrator this kernel is being used for
-     * @param forcesAreValid if the context has been modified since the last time step, this will be
-     *                       false to show that cached forces are invalid and must be recalculated.
-     *                       On exit, this should specify whether the cached forces are valid at the
-     *                       end of the step.
-     */
-    double computeKineticEnergy(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid);
-    /**
-     * Get the values of all global variables.
-     *
-     * @param context   the context in which to execute this kernel
-     * @param values    on exit, this contains the values
-     */
-    void getGlobalVariables(ContextImpl& context, std::vector<double>& values) const;
-    /**
-     * Set the values of all global variables.
-     *
-     * @param context   the context in which to execute this kernel
-     * @param values    a vector containing the values
-     */
-    void setGlobalVariables(ContextImpl& context, const std::vector<double>& values);
-    /**
-     * Get the values of a per-DOF variable.
-     *
-     * @param context   the context in which to execute this kernel
-     * @param variable  the index of the variable to get
-     * @param values    on exit, this contains the values
-     */
-    void getPerDofVariable(ContextImpl& context, int variable, std::vector<Vec3>& values) const;
-    /**
-     * Set the values of a per-DOF variable.
-     *
-     * @param context   the context in which to execute this kernel
-     * @param variable  the index of the variable to get
-     * @param values    a vector containing the values
-     */
-    void setPerDofVariable(ContextImpl& context, int variable, const std::vector<Vec3>& values);
-private:
-    class ReorderListener;
-    class GlobalTarget;
-    class DerivFunction;
-    std::string createPerDofComputation(const std::string& variable, const Lepton::ParsedExpression& expr, CustomIntegrator& integrator,
-        const std::string& forceName, const std::string& energyName, std::vector<const TabulatedFunction*>& functions,
-        std::vector<std::pair<std::string, std::string> >& functionNames);
-    void prepareForComputation(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid);
-    Lepton::ExpressionTreeNode replaceDerivFunctions(const Lepton::ExpressionTreeNode& node, OpenMM::ContextImpl& context);
-    void findExpressionsForDerivs(const Lepton::ExpressionTreeNode& node, std::vector<std::pair<Lepton::ExpressionTreeNode, std::string> >& variableNodes);
-    void recordGlobalValue(double value, GlobalTarget target, CustomIntegrator& integrator);
-    void recordChangedParameters(ContextImpl& context);
-    bool evaluateCondition(int step);
-    ComputeContext& cc;
-    double energy;
-    float energyFloat;
-    int numGlobalVariables, sumWorkGroupSize;
-    bool hasInitializedKernels, deviceGlobalsAreCurrent, modifiesParameters, hasAnyConstraints, needsEnergyParamDerivs;
-    std::vector<bool> deviceValuesAreCurrent;
-    mutable std::vector<bool> localValuesAreCurrent;
-    ComputeArray globalValues, sumBuffer, summedValue;
-    ComputeArray uniformRandoms, randomSeed, perDofEnergyParamDerivs;
-    std::vector<ComputeArray> tabulatedFunctions, perDofValues;
-    std::map<int, double> savedEnergy;
-    std::map<int, ComputeArray> savedForces;
-    std::set<int> validSavedForces;
-    mutable std::vector<std::vector<mm_float4> > localPerDofValuesFloat;
-    mutable std::vector<std::vector<mm_double4> > localPerDofValuesDouble;
-    std::map<std::string, double> energyParamDerivs;
-    std::vector<std::string> perDofEnergyParamDerivNames;
-    std::vector<double> localPerDofEnergyParamDerivs;
-    std::vector<double> localGlobalValues;
-    std::vector<double> initialGlobalVariables;
-    std::vector<std::vector<ComputeKernel> > kernels;
-    ComputeKernel randomKernel, kineticEnergyKernel, sumKineticEnergyKernel;
-    std::vector<CustomIntegrator::ComputationType> stepType;
-    std::vector<CustomIntegratorUtilities::Comparison> comparisons;
-    std::vector<std::vector<Lepton::CompiledExpression> > globalExpressions;
-    CompiledExpressionSet expressionSet;
-    std::vector<bool> needsGlobals, needsForces, needsEnergy;
-    std::vector<bool> computeBothForceAndEnergy, invalidatesForces, merged;
-    std::vector<int> forceGroupFlags, blockEnd, requiredGaussian, requiredUniform;
-    std::vector<int> stepEnergyVariableIndex, globalVariableIndex, parameterVariableIndex;
-    int gaussianVariableIndex, uniformVariableIndex, dtVariableIndex;
-    std::vector<std::string> parameterNames;
-    std::vector<GlobalTarget> stepTarget;
-};
-
-class CommonIntegrateCustomStepKernel::GlobalTarget {
-public:
-    CommonIntegrateCustomStepKernel::GlobalTargetType type;
-    int variableIndex;
-    GlobalTarget() {
-    }
-    GlobalTarget(CommonIntegrateCustomStepKernel::GlobalTargetType type, int variableIndex) : type(type), variableIndex(variableIndex) {
-    }
 };
 
 /**
