+++ platforms/common/src/CommonKernels.cpp
@@ -2996,677 +2996,6 @@ double CommonIntegrateLangevinMiddleStepKernel::computeKineticEnergy(ContextImpl
     return cc.getIntegrationUtilities().computeKineticEnergy(0.0);
 }
 
-void CommonIntegrateNoseHooverStepKernel::initialize(const System& system, const NoseHooverIntegrator& integrator) {
-    cc.initializeContexts();
-    ContextSelector selector(cc);
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    map<string, string> defines;
-    defines["BOLTZ"] = cc.doubleToString(BOLTZ, true);
-    ComputeProgram program = cc.compileProgram(CommonKernelSources::noseHooverIntegrator, defines);
-    kernel1 = program->createKernel("integrateNoseHooverMiddlePart1");
-    kernel2 = program->createKernel("integrateNoseHooverMiddlePart2");
-    kernel3 = program->createKernel("integrateNoseHooverMiddlePart3");
-    kernel4 = program->createKernel("integrateNoseHooverMiddlePart4");
-    if (useDouble) {
-        oldDelta.initialize<mm_double4>(cc, cc.getPaddedNumAtoms(), "oldDelta");
-    } else {
-        oldDelta.initialize<mm_float4>(cc, cc.getPaddedNumAtoms(), "oldDelta");
-    }
-    kernelHardWall = program->createKernel("integrateNoseHooverHardWall");
-    prevMaxPairDistance = -1.0f;
-    maxPairDistanceBuffer.initialize<float>(cc, 1, "maxPairDistanceBuffer");
-
-    int workGroupSize = std::min(cc.getMaxThreadBlockSize(), 512);
-    defines["WORK_GROUP_SIZE"] = std::to_string(workGroupSize);
-
-    defines["BEGIN_YS_LOOP"] = "const real arr[1] = {1.0};"
-                               "for(int i=0;i<1;++i) {"
-                               "const real ys = arr[i];";
-    defines["END_YS_LOOP"] = "}";
-    program = cc.compileProgram(CommonKernelSources::noseHooverChain, defines);
-    propagateKernels[1] = program->createKernel("propagateNoseHooverChain");
-
-    defines["BEGIN_YS_LOOP"] = "const real arr[3] = {0.828981543588751, -0.657963087177502, 0.828981543588751};"
-                               "for(int i=0;i<3;++i) {"
-                               "const real ys = arr[i];";
-    program = cc.compileProgram(CommonKernelSources::noseHooverChain, defines);
-    propagateKernels[3] = program->createKernel("propagateNoseHooverChain");
-
-    defines["BEGIN_YS_LOOP"] = "const real arr[5] = {0.2967324292201065, 0.2967324292201065, -0.186929716880426, 0.2967324292201065, 0.2967324292201065};"
-                               "for(int i=0;i<5;++i) {"
-                               "const real ys = arr[i];";
-    program = cc.compileProgram(CommonKernelSources::noseHooverChain, defines);
-    propagateKernels[5] = program->createKernel("propagateNoseHooverChain");
-
-    defines["BEGIN_YS_LOOP"] = "const real arr[7] = {0.784513610477560, 0.235573213359357, -1.17767998417887, 1.31518632068391,-1.17767998417887, 0.235573213359357, 0.784513610477560};"
-                               "for(int i=0;i<7;++i) {"
-                               "const real ys = arr[i];";
-    program = cc.compileProgram(CommonKernelSources::noseHooverChain, defines);
-    propagateKernels[7] = program->createKernel("propagateNoseHooverChain");
-    program = cc.compileProgram(CommonKernelSources::noseHooverChain, defines);
-    reduceEnergyKernel = program->createKernel("reduceEnergyPair");
-
-    computeHeatBathEnergyKernel = program->createKernel("computeHeatBathEnergy");
-    computeAtomsKineticEnergyKernel = program->createKernel("computeAtomsKineticEnergy");
-    computePairsKineticEnergyKernel = program->createKernel("computePairsKineticEnergy");
-    scaleAtomsVelocitiesKernel = program->createKernel("scaleAtomsVelocities");
-    scalePairsVelocitiesKernel = program->createKernel("scalePairsVelocities");
-    int energyBufferSize = cc.getEnergyBuffer().getSize();
-    if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision())
-        energyBuffer.initialize<mm_double2>(cc, energyBufferSize, "energyBuffer");
-    else
-        energyBuffer.initialize<mm_float2>(cc, energyBufferSize, "energyBuffer");
-}
-
-void CommonIntegrateNoseHooverStepKernel::execute(ContextImpl& context, const NoseHooverIntegrator& integrator) {
-    ContextSelector selector(cc);
-    IntegrationUtilities& integration = cc.getIntegrationUtilities();
-    int paddedNumAtoms = cc.getPaddedNumAtoms();
-    double dt = integrator.getStepSize();
-    cc.getIntegrationUtilities().setNextStepSize(dt);
-
-    // If the atom reordering has occured, the forces from the previous step are permuted and thus invalid.
-    // They need to be either sorted or recomputed; here we choose the latter.
-    if (cc.getAtomsWereReordered())
-        context.calcForcesAndEnergy(true, false, integrator.getIntegrationForceGroups());
-
-    const auto& atomList = integrator.getAllThermostatedIndividualParticles();
-    const auto& pairList = integrator.getAllThermostatedPairs();
-    int numAtoms = atomList.size();
-    int numPairs = pairList.size();
-    int numParticles = numAtoms + 2*numPairs;
-    float maxPairDistance = integrator.getMaximumPairDistance();
-    // Make sure atom and pair metadata is uploaded and has the correct dimensions
-    if (prevMaxPairDistance != maxPairDistance) {
-        std::vector<float> tmp(1, maxPairDistance);
-        maxPairDistanceBuffer.upload(tmp);
-        prevMaxPairDistance = maxPairDistance;
-    }
-    if (numAtoms !=0 && (!atomListBuffer.isInitialized() || atomListBuffer.getSize() != numAtoms)) {
-        if (atomListBuffer.isInitialized())
-            atomListBuffer.resize(atomList.size());
-        else
-            atomListBuffer.initialize<int>(cc, atomList.size(), "atomListBuffer");
-        atomListBuffer.upload(atomList);
-    }
-    if (numPairs !=0 && (!pairListBuffer.isInitialized() || pairListBuffer.getSize() != numPairs)) {
-        if (pairListBuffer.isInitialized()) {
-            pairListBuffer.resize(pairList.size());
-            pairTemperatureBuffer.resize(pairList.size());
-        }
-        else {
-            pairListBuffer.initialize<mm_int2>(cc, pairList.size(), "pairListBuffer");
-            pairTemperatureBuffer.initialize<float>(cc, pairList.size(), "pairTemperatureBuffer");
-        }
-        std::vector<mm_int2> tmp;
-        std::vector<float> tmp2;
-        for(const auto &pair : pairList) {
-            tmp.push_back(mm_int2(std::get<0>(pair), std::get<1>(pair)));
-            tmp2.push_back(std::get<2>(pair));
-        }
-        pairListBuffer.upload(tmp);
-        pairTemperatureBuffer.upload(tmp2);
-    }
-    int totalAtoms = cc.getNumAtoms();
-    if (!hasInitializedKernels) {
-        hasInitializedKernels = true;
-        kernel1->addArg(numAtoms);
-        kernel1->addArg(numPairs);
-        kernel1->addArg(paddedNumAtoms);
-        kernel1->addArg(cc.getVelm());
-        kernel1->addArg(cc.getLongForceBuffer());
-        kernel1->addArg(integration.getStepSize());
-        kernel1->addArg(numAtoms > 0 ? atomListBuffer : cc.getEnergyBuffer()); // The array is not used if num == 0
-        kernel1->addArg(numPairs > 0 ? pairListBuffer : cc.getEnergyBuffer()); // The array is not used if num == 0
-        kernel2->addArg(totalAtoms);
-        kernel2->addArg(cc.getVelm());
-        kernel2->addArg(integration.getPosDelta());
-        kernel2->addArg(oldDelta);
-        kernel2->addArg(integration.getStepSize());
-        kernel3->addArg(totalAtoms);
-        kernel3->addArg(cc.getVelm());
-        kernel3->addArg(integration.getPosDelta());
-        kernel3->addArg(oldDelta);
-        kernel3->addArg(integration.getStepSize());
-        kernel4->addArg(totalAtoms);
-        kernel4->addArg(cc.getPosq());
-        kernel4->addArg(cc.getVelm());
-        kernel4->addArg(integration.getPosDelta());
-        kernel4->addArg(oldDelta);
-        kernel4->addArg(integration.getStepSize());
-        if (cc.getUseMixedPrecision())
-            kernel4->addArg(cc.getPosqCorrection());
-        if (numPairs > 0) {
-            kernelHardWall->addArg(numPairs);
-            kernelHardWall->addArg(maxPairDistanceBuffer);
-            kernelHardWall->addArg(integration.getStepSize());
-            kernelHardWall->addArg(cc.getPosq());
-            kernelHardWall->addArg(cc.getVelm());
-            kernelHardWall->addArg(pairListBuffer);
-            kernelHardWall->addArg(pairTemperatureBuffer);
-            if (cc.getUseMixedPrecision())
-                kernelHardWall->addArg(cc.getPosqCorrection());
-        }
-    }
-
-    /*
-     * Carry out the LF-middle integration (c.f. J. Phys. Chem. A 2019, 123, 6056−6079)
-     */
-    // Velocity update
-    kernel1->execute(std::max(numAtoms, numPairs));
-    integration.applyVelocityConstraints(integrator.getConstraintTolerance());
-    // Position update
-    kernel2->execute(numParticles);
-    // Apply the thermostat
-    int numChains = integrator.getNumThermostats();
-    for(int chain = 0; chain < numChains; ++chain) {
-        const auto &thermostatChain = integrator.getThermostat(chain);
-        auto KEs = computeMaskedKineticEnergy(context, thermostatChain, false);
-        auto scaleFactors = propagateChain(context, thermostatChain, KEs, dt);
-        scaleVelocities(context, thermostatChain, scaleFactors);
-    }
-    // Position update
-    kernel3->execute(numParticles);
-    integration.applyConstraints(integrator.getConstraintTolerance());
-    // Apply constraint forces
-    kernel4->execute(numAtoms);
-    // Make sure any Drude-like particles have not wandered too far from home
-    if (numPairs > 0) kernelHardWall->execute(numPairs);
-    integration.computeVirtualSites();
-
-    // Update the time and step count.
-    cc.setTime(cc.getTime()+dt);
-    cc.setStepCount(cc.getStepCount()+1);
-    cc.reorderAtoms();
-
-    // Reduce UI lag.
-
-    flushPeriodically(cc);
-}
-
-double CommonIntegrateNoseHooverStepKernel::computeKineticEnergy(ContextImpl& context, const NoseHooverIntegrator& integrator) {
-    return cc.getIntegrationUtilities().computeKineticEnergy(0);
-}
-
-
-std::pair<double, double> CommonIntegrateNoseHooverStepKernel::propagateChain(ContextImpl& context, const NoseHooverChain &nhc, std::pair<double, double> kineticEnergies, double timeStep) {
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    int chainID = nhc.getChainID();
-    int nAtoms = nhc.getThermostatedAtoms().size();
-    int nPairs = nhc.getThermostatedPairs().size();
-    int chainLength = nhc.getChainLength();
-    int numYS = nhc.getNumYoshidaSuzukiTimeSteps();
-    int numMTS = nhc.getNumMultiTimeSteps();
-
-    if (numYS != 1 && numYS != 3 && numYS != 5 && numYS != 7) {
-        throw OpenMMException("Number of Yoshida Suzuki time steps has to be 1, 3, 5, or 7.");
-    }
-
-    if (!scaleFactorBuffer.isInitialized() || scaleFactorBuffer.getSize() == 0) {
-        if (useDouble) {
-            std::vector<mm_double2> zeros{{0,0}};
-            if (scaleFactorBuffer.isInitialized())
-                scaleFactorBuffer.resize(1);
-            else
-                scaleFactorBuffer.initialize<mm_double2>(cc, 1, "scaleFactorBuffer");
-            scaleFactorBuffer.upload(zeros);
-        }
-        else {
-            std::vector<mm_float2> zeros{{0,0}};
-            if (scaleFactorBuffer.isInitialized())
-                scaleFactorBuffer.resize(1);
-            else
-                scaleFactorBuffer.initialize<mm_float2>(cc, 1, "scaleFactorBuffer");
-            scaleFactorBuffer.upload(zeros);
-        }
-    }
-    if (!chainForces.isInitialized() || !chainMasses.isInitialized()) {
-        if (useDouble) {
-            std::vector<double> zeros(chainLength,0);
-            if (chainForces.isInitialized()) {
-                chainMasses.resize(chainLength);
-                chainForces.resize(chainLength);
-            }
-            else {
-                chainMasses.initialize<double>(cc, chainLength, "chainMasses");
-                chainForces.initialize<double>(cc, chainLength, "chainForces");
-            }
-            chainMasses.upload(zeros);
-            chainForces.upload(zeros);
-        }
-        else {
-            std::vector<float> zeros(chainLength,0);
-            if (chainForces.isInitialized()) {
-                chainMasses.resize(chainLength);
-                chainForces.resize(chainLength);
-            }
-            else {
-                chainMasses.initialize<float>(cc, chainLength, "chainMasses");
-                chainForces.initialize<float>(cc, chainLength, "chainForces");
-            }
-            chainMasses.upload(zeros);
-            chainForces.upload(zeros);
-        }
-    }
-    if (chainForces.getSize() < chainLength)
-        chainForces.resize(chainLength);
-    if (chainMasses.getSize() < chainLength)
-        chainMasses.resize(chainLength);
-
-
-    // N.B. We ignore the incoming kineticEnergy and grab it from the device buffer instead
-    if (nAtoms) {
-        if (!chainState.count(2*chainID))
-            chainState[2*chainID] = ComputeArray();
-        if (!chainState.at(2*chainID).isInitialized() || chainState.at(2*chainID).getSize() != chainLength) {
-            // We need to upload the Common array
-            if (useDouble) {
-                if (chainState.at(2*chainID).isInitialized())
-                    chainState.at(2*chainID).resize(chainLength);
-                else
-                    chainState.at(2*chainID).initialize<mm_double2>(cc, chainLength, "chainState" + std::to_string(2*chainID));
-                std::vector<mm_double2> zeros(chainLength, mm_double2(0.0, 0.0));
-                chainState.at(2*chainID).upload(zeros.data());
-            }
-            else {
-                if (chainState.at(2*chainID).isInitialized())
-                    chainState.at(2*chainID).resize(chainLength);
-                else
-                    chainState.at(2*chainID).initialize<mm_float2>(cc, chainLength, "chainState" + std::to_string(2*chainID));
-                std::vector<mm_float2> zeros(chainLength, mm_float2(0.0f, 0.0f));
-                chainState.at(2*chainID).upload(zeros.data());
-            }
-        }
-    }
-
-    if (nPairs) {
-        if (!chainState.count(2*chainID+1))
-            chainState[2*chainID+1] = ComputeArray();
-        if (!chainState.at(2*chainID+1).isInitialized() || chainState.at(2*chainID+1).getSize() != chainLength) {
-            // We need to upload the Common array
-            if (useDouble) {
-                if (chainState.at(2*chainID+1).isInitialized())
-                    chainState.at(2*chainID+1).resize(chainLength);
-                else
-                    chainState.at(2*chainID+1).initialize<mm_double2>(cc, chainLength, "chainState" + std::to_string(2*chainID+1));
-                std::vector<mm_double2> zeros(chainLength, mm_double2(0.0, 0.0));
-                chainState.at(2*chainID+1).upload(zeros.data());
-            }
-            else {
-                if (chainState.at(2*chainID+1).isInitialized())
-                    chainState.at(2*chainID+1).resize(chainLength);
-                else
-                    chainState.at(2*chainID+1).initialize<mm_float2>(cc, chainLength, "chainState" + std::to_string(2*chainID+1));
-                std::vector<mm_float2> zeros(chainLength, mm_float2(0.0f, 0.0f));
-                chainState.at(2*chainID+1).upload(zeros.data());
-            }
-        }
-    }
-
-    if (!hasInitializedPropagateKernel) {
-        hasInitializedPropagateKernel = true;
-        propagateKernels[numYS]->addArg(); // ChainState
-        propagateKernels[numYS]->addArg(kineticEnergyBuffer);
-        propagateKernels[numYS]->addArg(scaleFactorBuffer);
-        propagateKernels[numYS]->addArg(chainMasses);
-        propagateKernels[numYS]->addArg(chainForces);
-        propagateKernels[numYS]->addArg(); // ChainType
-        propagateKernels[numYS]->addArg(chainLength);
-        propagateKernels[numYS]->addArg(numMTS);
-        propagateKernels[numYS]->addArg(); // numDoFs
-        propagateKernels[numYS]->addArg((float)timeStep);
-        propagateKernels[numYS]->addArg(); // kT
-        propagateKernels[numYS]->addArg(); // frequency
-    }
-
-    if (nAtoms) {
-        int chainType = 0;
-        double temperature = nhc.getTemperature();
-        float frequency = nhc.getCollisionFrequency();
-        double kT = BOLTZ * temperature;
-        int numDOFs = nhc.getNumDegreesOfFreedom();
-        propagateKernels[numYS]->setArg(0, chainState[2*chainID]);
-        propagateKernels[numYS]->setArg(5, chainType);
-        propagateKernels[numYS]->setArg(8, numDOFs);
-        if (useDouble) {
-            propagateKernels[numYS]->setArg(10, kT);
-        } else {
-            propagateKernels[numYS]->setArg(10, (float)kT);
-        }
-        propagateKernels[numYS]->setArg(11, frequency);
-        propagateKernels[numYS]->execute(1, 1);
-    }
-    if (nPairs) {
-        int chainType = 1;
-        double relativeTemperature = nhc.getRelativeTemperature();
-        float relativeFrequency = nhc.getRelativeCollisionFrequency();
-        double kT = BOLTZ * relativeTemperature;
-        int ndf = 3*nPairs;
-        propagateKernels[numYS]->setArg(0, chainState[2*chainID+1]);
-        propagateKernels[numYS]->setArg(5, chainType);
-        propagateKernels[numYS]->setArg(8, ndf);
-        if (useDouble) {
-            propagateKernels[numYS]->setArg(10, kT);
-        } else {
-            propagateKernels[numYS]->setArg(10, (float)kT);
-        }
-        propagateKernels[numYS]->setArg(11, relativeFrequency);
-        propagateKernels[numYS]->execute(1, 1);
-    }
-    return {0, 0};
-}
-
-double CommonIntegrateNoseHooverStepKernel::computeHeatBathEnergy(ContextImpl& context, const NoseHooverChain &nhc) {
-    ContextSelector selector(cc);
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    int chainID = nhc.getChainID();
-    int chainLength = nhc.getChainLength();
-
-    bool absChainIsValid = chainState.count(2*chainID) != 0 &&
-                           chainState[2*chainID].isInitialized() &&
-                           chainState[2*chainID].getSize() == chainLength;
-    bool relChainIsValid = chainState.count(2*chainID+1) != 0 &&
-                           chainState[2*chainID+1].isInitialized() &&
-                           chainState[2*chainID+1].getSize() == chainLength;
-
-    if (!absChainIsValid && !relChainIsValid) return 0.0;
-
-    if (!heatBathEnergy.isInitialized() || heatBathEnergy.getSize() == 0) {
-        if (useDouble) {
-            std::vector<double> one(1);
-            heatBathEnergy.initialize<double>(cc, 1, "heatBathEnergy");
-            heatBathEnergy.upload(one);
-        }
-        else {
-            std::vector<float> one(1);
-            heatBathEnergy.initialize<float>(cc, 1, "heatBathEnergy");
-            heatBathEnergy.upload(one);
-        }
-    }
-
-    cc.clearBuffer(heatBathEnergy);
-
-    if(!hasInitializedHeatBathEnergyKernel) {
-        hasInitializedHeatBathEnergyKernel = true;
-        computeHeatBathEnergyKernel->addArg(heatBathEnergy);
-        computeHeatBathEnergyKernel->addArg(chainLength);
-        computeHeatBathEnergyKernel->addArg(); // numDOFs
-        computeHeatBathEnergyKernel->addArg(); // kT
-        computeHeatBathEnergyKernel->addArg(); // frequency
-        computeHeatBathEnergyKernel->addArg(); // chainstate
-    }
-
-    if (absChainIsValid) {
-        int numDOFs = nhc.getNumDegreesOfFreedom();
-        double temperature = nhc.getTemperature();
-        float frequency = nhc.getCollisionFrequency();
-        double kT = BOLTZ * temperature;
-
-        computeHeatBathEnergyKernel->setArg(2, numDOFs);
-        if (useDouble) {
-            computeHeatBathEnergyKernel->setArg(3, kT);
-        } else {
-            computeHeatBathEnergyKernel->setArg(3, (float)kT);
-        }
-        computeHeatBathEnergyKernel->setArg(4, frequency);
-        computeHeatBathEnergyKernel->setArg(5, chainState[2*chainID]);
-        computeHeatBathEnergyKernel->execute(1, 1);
-    }
-    if (relChainIsValid) {
-        int numDOFs = 3 * nhc.getThermostatedPairs().size();
-        double temperature = nhc.getRelativeTemperature();
-        float frequency = nhc.getRelativeCollisionFrequency();
-        double kT = BOLTZ * temperature;
-
-        computeHeatBathEnergyKernel->setArg(2, numDOFs);
-        if (useDouble) {
-            computeHeatBathEnergyKernel->setArg(3, kT);
-        } else {
-            computeHeatBathEnergyKernel->setArg(3, (float)kT);
-        }
-        computeHeatBathEnergyKernel->setArg(4, frequency);
-        computeHeatBathEnergyKernel->setArg(5, chainState[2*chainID+1]);
-        computeHeatBathEnergyKernel->execute(1, 1);
-    }
-
-
-    void * pinnedBuffer = cc.getPinnedBuffer();
-    heatBathEnergy.download(pinnedBuffer);
-    if (useDouble)
-        return *((double*) pinnedBuffer);
-    else
-        return *((float*) pinnedBuffer);
-}
-
-std::pair<double, double> CommonIntegrateNoseHooverStepKernel::computeMaskedKineticEnergy(ContextImpl& context, const NoseHooverChain &nhc, bool downloadValue) {
-    ContextSelector selector(cc);
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    int chainID = nhc.getChainID();
-    const auto & nhcAtoms = nhc.getThermostatedAtoms();
-    const auto & nhcPairs = nhc.getThermostatedPairs();
-    int nAtoms = nhcAtoms.size();
-    int nPairs = nhcPairs.size();
-    if (nAtoms) {
-        if (!atomlists.count(chainID)) {
-            // We need to upload the Common array
-            atomlists[chainID] = ComputeArray();
-            atomlists[chainID].initialize<int>(cc, nAtoms, "atomlist" + std::to_string(chainID));
-            atomlists[chainID].upload(nhcAtoms);
-        }
-        if (atomlists[chainID].getSize() != nAtoms) {
-            throw OpenMMException("Number of atoms changed. Cannot be handled by the same Nose-Hoover thermostat.");
-        }
-    }
-    if (nPairs) {
-        if (!pairlists.count(chainID)) {
-            // We need to upload the Common array
-            pairlists[chainID] = ComputeArray();
-            pairlists[chainID].initialize<mm_int2>(cc, nPairs, "pairlist" + std::to_string(chainID));
-            std::vector<mm_int2> int2vec;
-            for(const auto &p : nhcPairs) int2vec.push_back(mm_int2(p.first, p.second));
-            pairlists[chainID].upload(int2vec);
-        }
-        if (pairlists[chainID].getSize() != nPairs) {
-            throw OpenMMException("Number of thermostated pairs changed. Cannot be handled by the same Nose-Hoover thermostat.");
-        }
-    }
-    if (!kineticEnergyBuffer.isInitialized() || kineticEnergyBuffer.getSize() == 0) {
-        if (useDouble) {
-            std::vector<mm_double2> zeros{{0,0}};
-            kineticEnergyBuffer.initialize<mm_double2>(cc, 1, "kineticEnergyBuffer");
-            kineticEnergyBuffer.upload(zeros);
-        }
-        else {
-            std::vector<mm_float2> zeros{{0,0}};
-            kineticEnergyBuffer.initialize<mm_float2>(cc, 1, "kineticEnergyBuffer");
-            kineticEnergyBuffer.upload(zeros);
-        }
-    }
-
-    int workGroupSize = std::min(cc.getMaxThreadBlockSize(), 512);
-    if (!hasInitializedKineticEnergyKernel) {
-        hasInitializedKineticEnergyKernel = true;
-        computeAtomsKineticEnergyKernel->addArg(energyBuffer);
-        computeAtomsKineticEnergyKernel->addArg(); // nAtoms
-        computeAtomsKineticEnergyKernel->addArg(cc.getVelm());
-        computeAtomsKineticEnergyKernel->addArg(); // atom list
-
-        computePairsKineticEnergyKernel->addArg(energyBuffer);
-        computePairsKineticEnergyKernel->addArg(); // nPairs
-        computePairsKineticEnergyKernel->addArg(cc.getVelm());
-        computePairsKineticEnergyKernel->addArg(); // pair list
-
-        reduceEnergyKernel->addArg(energyBuffer);
-        reduceEnergyKernel->addArg(kineticEnergyBuffer);
-        reduceEnergyKernel->addArg((int) energyBuffer.getSize());
-    }
-
-    cc.clearBuffer(energyBuffer);
-    if (nAtoms) {
-        computeAtomsKineticEnergyKernel->setArg(1, nAtoms);
-        computeAtomsKineticEnergyKernel->setArg(3, atomlists[chainID]);
-        computeAtomsKineticEnergyKernel->execute(nAtoms);
-    }
-    if (nPairs) {
-        computePairsKineticEnergyKernel->setArg(1, nPairs);
-        computePairsKineticEnergyKernel->setArg(3, pairlists[chainID]);
-        computePairsKineticEnergyKernel->execute(nPairs);
-    }
-    reduceEnergyKernel->execute(workGroupSize, workGroupSize);
-
-    std::pair<double, double> KEs = {0, 0};
-    if (downloadValue) {
-        if (useDouble) {
-            mm_double2 tmp;
-            kineticEnergyBuffer.download(&tmp);
-            KEs.first = tmp.x;
-            KEs.second = tmp.y;
-        }
-        else {
-            mm_float2 tmp;
-            kineticEnergyBuffer.download(&tmp);
-            KEs.first = tmp.x;
-            KEs.second = tmp.y;
-        }
-    }
-    return KEs;
-}
-
-void CommonIntegrateNoseHooverStepKernel::scaleVelocities(ContextImpl& context, const NoseHooverChain &nhc, std::pair<double, double> scaleFactor) {
-    // For now we assume that the atoms and pairs info is valid, because compute{Atoms|Pairs}KineticEnergy must have been
-    // called before this kernel.  If that ever ceases to be true, some sanity checks are needed here.
-
-    int chainID = nhc.getChainID();
-    int nAtoms = nhc.getThermostatedAtoms().size();
-    int nPairs = nhc.getThermostatedPairs().size();
-    if (!hasInitializedScaleVelocitiesKernel) {
-        hasInitializedScaleVelocitiesKernel = true;
-        scaleAtomsVelocitiesKernel->addArg(scaleFactorBuffer);
-        scaleAtomsVelocitiesKernel->addArg(); // nAtoms
-        scaleAtomsVelocitiesKernel->addArg(cc.getVelm());
-        scaleAtomsVelocitiesKernel->addArg(); // atom list
-
-        scalePairsVelocitiesKernel->addArg(scaleFactorBuffer);
-        scalePairsVelocitiesKernel->addArg(); // nPairs
-        scalePairsVelocitiesKernel->addArg(cc.getVelm());
-        scalePairsVelocitiesKernel->addArg(); // pair list
-    }
-    if (nAtoms) {
-        scaleAtomsVelocitiesKernel->setArg(1, nAtoms);
-        scaleAtomsVelocitiesKernel->setArg(3, atomlists[chainID]);
-        scaleAtomsVelocitiesKernel->execute(nAtoms);
-    }
-    if (nPairs) {
-        scalePairsVelocitiesKernel->setArg(1, nPairs);
-        scalePairsVelocitiesKernel->setArg(3, pairlists[chainID]);
-        scalePairsVelocitiesKernel->execute(nPairs);
-    }
-}
-
-void CommonIntegrateNoseHooverStepKernel::createCheckpoint(ContextImpl& context, ostream& stream) const {
-    ContextSelector selector(cc);
-    int numChains = chainState.size();
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    stream.write((char*) &numChains, sizeof(int));
-    for (auto& state : chainState){
-        int chainID = state.first;
-        int chainLength = state.second.getSize();
-        stream.write((char*) &chainID, sizeof(int));
-        stream.write((char*) &chainLength, sizeof(int));
-        if (useDouble) {
-            vector<mm_double2> stateVec;
-            state.second.download(stateVec);
-            stream.write((char*) stateVec.data(), sizeof(mm_double2)*chainLength);
-        }
-        else {
-            vector<mm_float2> stateVec;
-            state.second.download(stateVec);
-            stream.write((char*) stateVec.data(), sizeof(mm_float2)*chainLength);
-        }
-    }
-}
-
-void CommonIntegrateNoseHooverStepKernel::loadCheckpoint(ContextImpl& context, istream& stream) {
-    ContextSelector selector(cc);
-    int numChains;
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    stream.read((char*) &numChains, sizeof(int));
-    chainState.clear();
-    for (int i = 0; i < numChains; i++) {
-        int chainID, chainLength;
-        stream.read((char*) &chainID, sizeof(int));
-        stream.read((char*) &chainLength, sizeof(int));
-        if (useDouble) {
-            chainState[chainID] = ComputeArray();
-            chainState[chainID].initialize<mm_double2>(cc, chainLength, "chainState" + to_string(chainID));
-            vector<mm_double2> stateVec(chainLength);
-            stream.read((char*) &stateVec[0], sizeof(mm_double2)*chainLength);
-            chainState[chainID].upload(stateVec);
-        }
-        else {
-            chainState[chainID] = ComputeArray();
-            chainState[chainID].initialize<mm_float2>(cc, chainLength, "chainState" + to_string(chainID));
-            vector<mm_float2> stateVec(chainLength);
-            stream.read((char*) &stateVec[0], sizeof(mm_float2)*chainLength);
-            chainState[chainID].upload(stateVec);
-        }
-    }
-}
-
-void CommonIntegrateNoseHooverStepKernel::getChainStates(ContextImpl& context, vector<vector<double> >& positions, vector<vector<double> >& velocities) const {
-    ContextSelector selector(cc);
-    int numChains = chainState.size();
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    positions.clear();
-    velocities.clear();
-    positions.resize(numChains);
-    velocities.resize(numChains);
-    for (int i = 0; i < numChains; i++) {
-        const ComputeArray& state = chainState.at(i);
-        if (useDouble) {
-            vector<mm_double2> stateVec;
-            state.download(stateVec);
-            for (int j = 0; j < stateVec.size(); j++) {
-                positions[i].push_back(stateVec[j].x);
-                velocities[i].push_back(stateVec[j].y);
-            }
-        }
-        else {
-            vector<mm_float2> stateVec;
-            state.download(stateVec);
-            for (int j = 0; j < stateVec.size(); j++) {
-                positions[i].push_back((float) stateVec[j].x);
-                velocities[i].push_back((float) stateVec[j].y);
-            }
-        }
-    }
-}
-
-void CommonIntegrateNoseHooverStepKernel::setChainStates(ContextImpl& context, const vector<vector<double> >& positions, const vector<vector<double> >& velocities) {
-    ContextSelector selector(cc);
-    int numChains = positions.size();
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    chainState.clear();
-    for (int i = 0; i < numChains; i++) {
-        int chainLength = positions[i].size();
-        chainState[i] = ComputeArray();
-        if (useDouble) {
-            chainState[i].initialize<mm_double2>(cc, chainLength, "chainState"+cc.intToString(i));
-            vector<mm_double2> stateVec;
-            for (int j = 0; j < chainLength; j++)
-                stateVec.push_back(mm_double2(positions[i][j], velocities[i][j]));
-            chainState[i].upload(stateVec);
-        }
-        else {
-            chainState[i].initialize<mm_float2>(cc, chainLength, "chainState"+cc.intToString(i));
-            vector<mm_float2> stateVec;
-            for (int j = 0; j < chainLength; j++)
-                stateVec.push_back(mm_float2((float) positions[i][j], (float) velocities[i][j]));
-            chainState[i].upload(stateVec);
-        }
-    }
-}
-
 void CommonIntegrateBrownianStepKernel::initialize(const System& system, const BrownianIntegrator& integrator) {
     cc.initializeContexts();
     ContextSelector selector(cc);
@@ -3962,905 +3291,6 @@ double CommonIntegrateVariableLangevinStepKernel::execute(ContextImpl& context, 
 
 double CommonIntegrateVariableLangevinStepKernel::computeKineticEnergy(ContextImpl& context, const VariableLangevinIntegrator& integrator) {
     return cc.getIntegrationUtilities().computeKineticEnergy(0.5*integrator.getStepSize());
-}
-
-class CommonIntegrateCustomStepKernel::ReorderListener : public ComputeContext::ReorderListener {
-public:
-    ReorderListener(ComputeContext& cc, vector<ComputeArray>& perDofValues, vector<vector<mm_float4> >& localPerDofValuesFloat, vector<vector<mm_double4> >& localPerDofValuesDouble, vector<bool>& deviceValuesAreCurrent) :
-            cc(cc), perDofValues(perDofValues), localPerDofValuesFloat(localPerDofValuesFloat), localPerDofValuesDouble(localPerDofValuesDouble), deviceValuesAreCurrent(deviceValuesAreCurrent) {
-        int numAtoms = cc.getNumAtoms();
-        lastAtomOrder.resize(numAtoms);
-        for (int i = 0; i < numAtoms; i++)
-            lastAtomOrder[i] = cc.getAtomIndex()[i];
-    }
-    void execute() {
-        // Reorder the per-DOF variables to reflect the new atom order.
-
-        if (perDofValues.size() == 0)
-            return;
-        int numAtoms = cc.getNumAtoms();
-        const vector<int>& order = cc.getAtomIndex();
-        for (int index = 0; index < perDofValues.size(); index++) {
-            if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision()) {
-                if (deviceValuesAreCurrent[index])
-                    perDofValues[index].download(localPerDofValuesDouble[index]);
-                vector<mm_double4> swap(numAtoms);
-                for (int i = 0; i < numAtoms; i++)
-                    swap[lastAtomOrder[i]] = localPerDofValuesDouble[index][i];
-                for (int i = 0; i < numAtoms; i++)
-                    localPerDofValuesDouble[index][i] = swap[order[i]];
-                perDofValues[index].upload(localPerDofValuesDouble[index]);
-            }
-            else {
-                if (deviceValuesAreCurrent[index])
-                    perDofValues[index].download(localPerDofValuesFloat[index]);
-                vector<mm_float4> swap(numAtoms);
-                for (int i = 0; i < numAtoms; i++)
-                    swap[lastAtomOrder[i]] = localPerDofValuesFloat[index][i];
-                for (int i = 0; i < numAtoms; i++)
-                    localPerDofValuesFloat[index][i] = swap[order[i]];
-                perDofValues[index].upload(localPerDofValuesFloat[index]);
-            }
-            deviceValuesAreCurrent[index] = true;
-        }
-        for (int i = 0; i < numAtoms; i++)
-            lastAtomOrder[i] = order[i];
-    }
-private:
-    ComputeContext& cc;
-    vector<ComputeArray>& perDofValues;
-    vector<vector<mm_float4> >& localPerDofValuesFloat;
-    vector<vector<mm_double4> >& localPerDofValuesDouble;
-    vector<bool>& deviceValuesAreCurrent;
-    vector<int> lastAtomOrder;
-};
-
-class CommonIntegrateCustomStepKernel::DerivFunction : public CustomFunction {
-public:
-    DerivFunction(map<string, double>& energyParamDerivs, const string& param) : energyParamDerivs(energyParamDerivs), param(param) {
-    }
-    int getNumArguments() const {
-        return 0;
-    }
-    double evaluate(const double* arguments) const {
-        return energyParamDerivs[param];
-    }
-    double evaluateDerivative(const double* arguments, const int* derivOrder) const {
-        return 0;
-    }
-    CustomFunction* clone() const {
-        return new DerivFunction(energyParamDerivs, param);
-    }
-private:
-    map<string, double>& energyParamDerivs;
-    string param;
-};
-
-void CommonIntegrateCustomStepKernel::initialize(const System& system, const CustomIntegrator& integrator) {
-    cc.initializeContexts();
-    ContextSelector selector(cc);
-    cc.getIntegrationUtilities().initRandomNumberGenerator(integrator.getRandomNumberSeed());
-    numGlobalVariables = integrator.getNumGlobalVariables();
-    int elementSize = (cc.getUseDoublePrecision() || cc.getUseMixedPrecision() ? sizeof(double) : sizeof(float));
-    sumBuffer.initialize(cc, system.getNumParticles(), elementSize, "sumBuffer");
-    summedValue.initialize(cc, 1, elementSize, "summedValue");
-    perDofValues.resize(integrator.getNumPerDofVariables());
-    localPerDofValuesFloat.resize(perDofValues.size());
-    localPerDofValuesDouble.resize(perDofValues.size());
-    for (int i = 0; i < perDofValues.size(); i++)
-        perDofValues[i].initialize(cc, system.getNumParticles(), 4*elementSize, "perDofVariables");
-    localValuesAreCurrent.resize(integrator.getNumPerDofVariables(), false);
-    deviceValuesAreCurrent.resize(integrator.getNumPerDofVariables(), false);
-    cc.addReorderListener(new ReorderListener(cc, perDofValues, localPerDofValuesFloat, localPerDofValuesDouble, deviceValuesAreCurrent));
-    SimTKOpenMMUtilities::setRandomNumberSeed(integrator.getRandomNumberSeed());
-}
-
-string CommonIntegrateCustomStepKernel::createPerDofComputation(const string& variable, const Lepton::ParsedExpression& expr, CustomIntegrator& integrator,
-        const string& forceName, const string& energyName, vector<const TabulatedFunction*>& functions, vector<pair<string, string> >& functionNames) {
-    string tempType = (cc.getSupportsDoublePrecision() ? "double3" : "float3");
-    map<string, Lepton::ParsedExpression> expressions;
-    expressions[tempType+" tempResult = "] = expr;
-    map<string, string> variables;
-    variables["x"] = "make_"+tempType+"(position.x, position.y, position.z)";
-    variables["v"] = "make_"+tempType+"(velocity.x, velocity.y, velocity.z)";
-    variables[forceName] = "make_"+tempType+"(f.x, f.y, f.z)";
-    variables["gaussian"] = "make_"+tempType+"(gaussian.x, gaussian.y, gaussian.z)";
-    variables["uniform"] = "make_"+tempType+"(uniform.x, uniform.y, uniform.z)";
-    variables["m"] = "mass";
-    variables["dt"] = "stepSize";
-    if (energyName != "")
-        variables[energyName] = "make_"+tempType+"(energy)";
-    for (int i = 0; i < integrator.getNumGlobalVariables(); i++)
-        variables[integrator.getGlobalVariableName(i)] = "make_"+tempType+"(globals["+cc.intToString(globalVariableIndex[i])+"])";
-    for (int i = 0; i < integrator.getNumPerDofVariables(); i++)
-        variables[integrator.getPerDofVariableName(i)] = "convertToTempType3(perDof"+cc.intToString(i)+")";
-    for (int i = 0; i < (int) parameterNames.size(); i++)
-        variables[parameterNames[i]] = "make_"+tempType+"(globals["+cc.intToString(parameterVariableIndex[i])+"])";
-    vector<pair<ExpressionTreeNode, string> > variableNodes;
-    findExpressionsForDerivs(expr.getRootNode(), variableNodes);
-    for (auto& var : variables)
-        variableNodes.push_back(make_pair(ExpressionTreeNode(new Operation::Variable(var.first)), var.second));
-    string result = cc.getExpressionUtilities().createExpressions(expressions, variableNodes, functions, functionNames, "temp", tempType);
-    if (variable == "x")
-        result += "position.x = tempResult.x; position.y = tempResult.y; position.z = tempResult.z;\n";
-    else if (variable == "v")
-        result += "velocity.x = tempResult.x; velocity.y = tempResult.y; velocity.z = tempResult.z;\n";
-    else if (variable == "")
-        result += "sum[index] = tempResult.x+tempResult.y+tempResult.z;\n";
-    else {
-        for (int i = 0; i < integrator.getNumPerDofVariables(); i++)
-            if (variable == integrator.getPerDofVariableName(i)) {
-                string varName = "perDof"+cc.intToString(i);
-                result += varName+".x = tempResult.x; "+varName+".y = tempResult.y; "+varName+".z = tempResult.z;\n";
-            }
-    }
-    return result;
-}
-
-void CommonIntegrateCustomStepKernel::prepareForComputation(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid) {
-    ContextSelector selector(cc);
-    IntegrationUtilities& integration = cc.getIntegrationUtilities();
-    int numAtoms = cc.getNumAtoms();
-    int numSteps = integrator.getNumComputations();
-    bool useDouble = cc.getUseDoublePrecision() || cc.getUseMixedPrecision();
-    string tempType = (cc.getSupportsDoublePrecision() ? "double3" : "float3");
-    string perDofType = (useDouble ? "double4" : "float4");
-    if (!hasInitializedKernels) {
-        hasInitializedKernels = true;
-        
-        // Initialize various data structures.
-        
-        const map<string, double>& params = context.getParameters();
-        for (auto& param : params)
-            parameterNames.push_back(param.first);
-        kernels.resize(integrator.getNumComputations());
-        requiredGaussian.resize(integrator.getNumComputations(), 0);
-        requiredUniform.resize(integrator.getNumComputations(), 0);
-        needsGlobals.resize(numSteps, false);
-        globalExpressions.resize(numSteps);
-        stepType.resize(numSteps);
-        stepTarget.resize(numSteps);
-        merged.resize(numSteps, false);
-        modifiesParameters = false;
-        sumWorkGroupSize = cc.getMaxThreadBlockSize();
-        if (sumWorkGroupSize > 512)
-            sumWorkGroupSize = 512;
-        map<string, string> defines;
-        defines["NUM_ATOMS"] = cc.intToString(cc.getNumAtoms());
-        defines["PADDED_NUM_ATOMS"] = cc.intToString(cc.getPaddedNumAtoms());
-        defines["WORK_GROUP_SIZE"] = cc.intToString(sumWorkGroupSize);
-
-        // Record the tabulated functions.
-
-        map<string, Lepton::CustomFunction*> functions;
-        vector<pair<string, string> > functionNames;
-        vector<const TabulatedFunction*> functionList;
-        vector<string> tableTypes;
-        tabulatedFunctions.resize(integrator.getNumTabulatedFunctions());
-        for (int i = 0; i < integrator.getNumTabulatedFunctions(); i++) {
-            functionList.push_back(&integrator.getTabulatedFunction(i));
-            string name = integrator.getTabulatedFunctionName(i);
-            string arrayName = "table"+cc.intToString(i);
-            functionNames.push_back(make_pair(name, arrayName));
-            functions[name] = createReferenceTabulatedFunction(integrator.getTabulatedFunction(i));
-            int width;
-            vector<float> f = cc.getExpressionUtilities().computeFunctionCoefficients(integrator.getTabulatedFunction(i), width);
-            tabulatedFunctions[i].initialize<float>(cc, f.size(), "TabulatedFunction");
-            tabulatedFunctions[i].upload(f);
-            if (width == 1)
-                tableTypes.push_back("float");
-            else
-                tableTypes.push_back("float"+cc.intToString(width));
-        }
-
-        // Record information about all the computation steps.
-
-        vector<string> variable(numSteps);
-        vector<int> forceGroup;
-        vector<vector<Lepton::ParsedExpression> > expression;
-        CustomIntegratorUtilities::analyzeComputations(context, integrator, expression, comparisons, blockEnd, invalidatesForces, needsForces, needsEnergy, computeBothForceAndEnergy, forceGroup, functions);
-        for (int step = 0; step < numSteps; step++) {
-            string expr;
-            integrator.getComputationStep(step, stepType[step], variable[step], expr);
-            if (stepType[step] == CustomIntegrator::WhileBlockStart)
-                blockEnd[blockEnd[step]] = step; // Record where to branch back to.
-            if (stepType[step] == CustomIntegrator::ComputeGlobal || stepType[step] == CustomIntegrator::IfBlockStart || stepType[step] == CustomIntegrator::WhileBlockStart)
-                for (auto& expr : expression[step])
-                    globalExpressions[step].push_back(ParsedExpression(replaceDerivFunctions(expr.getRootNode(), context)).createCompiledExpression());
-        }
-        for (int step = 0; step < numSteps; step++) {
-            for (auto& expr : globalExpressions[step])
-                expressionSet.registerExpression(expr);
-        }
-        
-        // Record the indices for variables in the CompiledExpressionSet.
-        
-        gaussianVariableIndex = expressionSet.getVariableIndex("gaussian");
-        uniformVariableIndex = expressionSet.getVariableIndex("uniform");
-        dtVariableIndex = expressionSet.getVariableIndex("dt");
-        for (int i = 0; i < integrator.getNumGlobalVariables(); i++)
-            globalVariableIndex.push_back(expressionSet.getVariableIndex(integrator.getGlobalVariableName(i)));
-        for (auto& name : parameterNames)
-            parameterVariableIndex.push_back(expressionSet.getVariableIndex(name));
-
-        // Record the variable names and flags for the force and energy in each step.
-
-        forceGroupFlags.resize(numSteps, integrator.getIntegrationForceGroups());
-        vector<string> forceGroupName;
-        vector<string> energyGroupName;
-        for (int i = 0; i < 32; i++) {
-            stringstream fname;
-            fname << "f" << i;
-            forceGroupName.push_back(fname.str());
-            stringstream ename;
-            ename << "energy" << i;
-            energyGroupName.push_back(ename.str());
-        }
-        vector<string> forceName(numSteps, "f");
-        vector<string> energyName(numSteps, "energy");
-        stepEnergyVariableIndex.resize(numSteps, expressionSet.getVariableIndex("energy"));
-        for (int step = 0; step < numSteps; step++) {
-            if (needsForces[step] && forceGroup[step] > -1)
-                forceName[step] = forceGroupName[forceGroup[step]];
-            if (needsEnergy[step] && forceGroup[step] > -1) {
-                energyName[step] = energyGroupName[forceGroup[step]];
-                stepEnergyVariableIndex[step] = expressionSet.getVariableIndex(energyName[step]);
-            }
-            if (forceGroup[step] > -1)
-                forceGroupFlags[step] = 1<<forceGroup[step];
-            if (forceGroupFlags[step] == -2 && step > 0)
-                forceGroupFlags[step] = forceGroupFlags[step-1];
-            if (forceGroupFlags[step] != -2 && savedForces.find(forceGroupFlags[step]) == savedForces.end()) {
-                savedForces[forceGroupFlags[step]] = ComputeArray();
-                savedForces[forceGroupFlags[step]].initialize(cc, cc.getLongForceBuffer().getSize(), cc.getLongForceBuffer().getElementSize(), "savedForces");
-            }
-        }
-        
-        // Allocate space for storing global values, both on the host and the device.
-        
-        localGlobalValues.resize(expressionSet.getNumVariables());
-        int elementSize = (cc.getUseDoublePrecision() || cc.getUseMixedPrecision() ? sizeof(double) : sizeof(float));
-        globalValues.initialize(cc, expressionSet.getNumVariables(), elementSize, "globalValues");
-        for (int i = 0; i < integrator.getNumGlobalVariables(); i++) {
-            localGlobalValues[globalVariableIndex[i]] = initialGlobalVariables[i];
-            expressionSet.setVariable(globalVariableIndex[i], initialGlobalVariables[i]);
-        }
-        for (int i = 0; i < (int) parameterVariableIndex.size(); i++) {
-            double value = context.getParameter(parameterNames[i]);
-            localGlobalValues[parameterVariableIndex[i]] = value;
-            expressionSet.setVariable(parameterVariableIndex[i], value);
-        }
-        int numContextParams = context.getParameters().size();
-        localPerDofEnergyParamDerivs.resize(numContextParams);
-        perDofEnergyParamDerivs.initialize(cc, max(1, numContextParams), elementSize, "perDofEnergyParamDerivs");
-        
-        // Record information about the targets of steps that will be stored in global variables.
-        
-        for (int step = 0; step < numSteps; step++) {
-            if (stepType[step] == CustomIntegrator::ComputeGlobal || stepType[step] == CustomIntegrator::ComputeSum) {
-                if (variable[step] == "dt")
-                    stepTarget[step].type = DT;
-                for (int i = 0; i < integrator.getNumGlobalVariables(); i++)
-                    if (variable[step] == integrator.getGlobalVariableName(i))
-                        stepTarget[step].type = VARIABLE;
-                for (auto& name : parameterNames)
-                    if (variable[step] == name) {
-                        stepTarget[step].type = PARAMETER;
-                        modifiesParameters = true;
-                    }
-                stepTarget[step].variableIndex = expressionSet.getVariableIndex(variable[step]);
-            }
-        }
-
-        // Identify which per-DOF steps are going to require global variables or context parameters.
-
-        for (int step = 0; step < numSteps; step++) {
-            if (stepType[step] == CustomIntegrator::ComputePerDof || stepType[step] == CustomIntegrator::ComputeSum) {
-                for (int i = 0; i < integrator.getNumGlobalVariables(); i++)
-                    if (usesVariable(expression[step][0], integrator.getGlobalVariableName(i)))
-                        needsGlobals[step] = true;
-                for (auto& name : parameterNames)
-                    if (usesVariable(expression[step][0], name))
-                        needsGlobals[step] = true;
-            }
-        }
-        
-        // Determine how each step will represent the position (as just a value, or a value plus a delta).
-        
-        hasAnyConstraints = (context.getSystem().getNumConstraints() > 0);
-        vector<bool> storePosAsDelta(numSteps, false);
-        vector<bool> loadPosAsDelta(numSteps, false);
-        if (hasAnyConstraints) {
-            bool beforeConstrain = false;
-            for (int step = numSteps-1; step >= 0; step--) {
-                if (stepType[step] == CustomIntegrator::ConstrainPositions)
-                    beforeConstrain = true;
-                else if (stepType[step] == CustomIntegrator::ComputePerDof && variable[step] == "x" && beforeConstrain)
-                    storePosAsDelta[step] = true;
-            }
-            bool storedAsDelta = false;
-            for (int step = 0; step < numSteps; step++) {
-                loadPosAsDelta[step] = storedAsDelta;
-                if (storePosAsDelta[step] == true)
-                    storedAsDelta = true;
-                if (stepType[step] == CustomIntegrator::ConstrainPositions)
-                    storedAsDelta = false;
-            }
-        }
-        
-        // Identify steps that can be merged into a single kernel.
-        
-        for (int step = 1; step < numSteps; step++) {
-            if (invalidatesForces[step-1] || forceGroupFlags[step] != forceGroupFlags[step-1])
-                continue;
-            if (stepType[step-1] == CustomIntegrator::ComputePerDof && stepType[step] == CustomIntegrator::ComputePerDof)
-                merged[step] = true;
-        }
-        for (int step = numSteps-1; step > 0; step--)
-            if (merged[step]) {
-                needsForces[step-1] = (needsForces[step] || needsForces[step-1]);
-                needsEnergy[step-1] = (needsEnergy[step] || needsEnergy[step-1]);
-                needsGlobals[step-1] = (needsGlobals[step] || needsGlobals[step-1]);
-                computeBothForceAndEnergy[step-1] = (computeBothForceAndEnergy[step] || computeBothForceAndEnergy[step-1]);
-            }
-        
-        // Loop over all steps and create the kernels for them.
-        
-        for (int step = 0; step < numSteps; step++) {
-            if ((stepType[step] == CustomIntegrator::ComputePerDof || stepType[step] == CustomIntegrator::ComputeSum) && !merged[step]) {
-                // Compute a per-DOF value.
-                
-                stringstream compute;
-                for (int i = 0; i < perDofValues.size(); i++)
-                    compute << tempType<<" perDof"<<cc.intToString(i)<<" = convertToTempType3(perDofValues"<<cc.intToString(i)<<"[index]);\n";
-                int numGaussian = 0, numUniform = 0;
-                for (int j = step; j < numSteps && (j == step || merged[j]); j++) {
-                    numGaussian += numAtoms*usesVariable(expression[j][0], "gaussian");
-                    numUniform += numAtoms*usesVariable(expression[j][0], "uniform");
-                    compute << "{\n";
-                    if (numGaussian > 0)
-                        compute << "float4 gaussian = gaussianValues[gaussianIndex+index];\n";
-                    if (numUniform > 0)
-                        compute << "float4 uniform = uniformValues[uniformIndex+index];\n";
-                    compute << createPerDofComputation(stepType[j] == CustomIntegrator::ComputePerDof ? variable[j] : "", expression[j][0], integrator, forceName[j], energyName[j], functionList, functionNames);
-                    if (variable[j] == "x") {
-                        if (storePosAsDelta[j]) {
-                            if (cc.getSupportsDoublePrecision())
-                                compute << "posDelta[index] = convertFromDouble4(position-loadPos(posq, posqCorrection, index));\n";
-                            else
-                                compute << "posDelta[index] = position-posq[index];\n";
-                        }
-                        else
-                            compute << "storePos(posq, posqCorrection, index, position);\n";
-                    }
-                    else if (variable[j] == "v") {
-                        if (cc.getSupportsDoublePrecision())
-                            compute << "velm[index] = convertFromDouble4(velocity);\n";
-                        else
-                            compute << "velm[index] = velocity;\n";
-                    }
-                    else {
-                        for (int i = 0; i < perDofValues.size(); i++)
-                            compute << "perDofValues"<<cc.intToString(i)<<"[index] = make_"<<perDofType<<"(perDof"<<cc.intToString(i)<<".x, perDof"<<cc.intToString(i)<<".y, perDof"<<cc.intToString(i)<<".z, 0);\n";
-                    }
-                    if (numGaussian > 0)
-                        compute << "gaussianIndex += NUM_ATOMS;\n";
-                    if (numUniform > 0)
-                        compute << "uniformIndex += NUM_ATOMS;\n";
-                    compute << "}\n";
-                }
-                map<string, string> replacements;
-                replacements["COMPUTE_STEP"] = compute.str();
-                stringstream args;
-                for (int i = 0; i < perDofValues.size(); i++) {
-                    string valueName = "perDofValues"+cc.intToString(i);
-                    args << ", GLOBAL " << perDofType << "* RESTRICT " << valueName;
-                }
-                for (int i = 0; i < (int) tableTypes.size(); i++)
-                    args << ", GLOBAL const " << tableTypes[i]<< "* RESTRICT table" << i;
-                replacements["PARAMETER_ARGUMENTS"] = args.str();
-                if (loadPosAsDelta[step])
-                    defines["LOAD_POS_AS_DELTA"] = "1";
-                else if (defines.find("LOAD_POS_AS_DELTA") != defines.end())
-                    defines.erase("LOAD_POS_AS_DELTA");
-                ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customIntegratorPerDof, replacements), defines);
-                ComputeKernel kernel = program->createKernel("computePerDof");
-                kernels[step].push_back(kernel);
-                requiredGaussian[step] = numGaussian;
-                requiredUniform[step] = numUniform;
-                kernel->addArg(cc.getPosq());
-                if (cc.getUseMixedPrecision())
-                    kernel->addArg(cc.getPosqCorrection());
-                else
-                    kernel->addArg(nullptr);
-                kernel->addArg(integration.getPosDelta());
-                kernel->addArg(cc.getVelm());
-                kernel->addArg(cc.getLongForceBuffer());
-                kernel->addArg(integration.getStepSize());
-                kernel->addArg(globalValues);
-                kernel->addArg(sumBuffer);
-                for (int i = 0; i < 4; i++)
-                    kernel->addArg();
-                kernel->addArg(perDofEnergyParamDerivs);
-                for (auto& array : perDofValues)
-                    kernel->addArg(array);
-                for (auto& array : tabulatedFunctions)
-                    kernel->addArg(array);
-                if (stepType[step] == CustomIntegrator::ComputeSum) {
-                    // Create a second kernel for this step that sums the values.
-
-                    program = cc.compileProgram(CommonKernelSources::customIntegrator, defines);
-                    kernel = program->createKernel(useDouble ? "computeDoubleSum" : "computeFloatSum");
-                    kernels[step].push_back(kernel);
-                    kernel->addArg(sumBuffer);
-                    kernel->addArg(summedValue);
-                    kernel->addArg(numAtoms);
-                }
-            }
-            else if (stepType[step] == CustomIntegrator::ConstrainPositions) {
-                // Apply position constraints.
-
-                ComputeProgram program = cc.compileProgram(CommonKernelSources::customIntegrator, defines);
-                ComputeKernel kernel = program->createKernel("applyPositionDeltas");
-                kernels[step].push_back(kernel);
-                kernel->addArg(cc.getPosq());
-                if (cc.getUseMixedPrecision())
-                    kernel->addArg(cc.getPosqCorrection());
-                else
-                    kernel->addArg(nullptr);
-                kernel->addArg(integration.getPosDelta());
-            }
-        }
-        
-        // Initialize the random number generator.
-        
-        int maxUniformRandoms = 1;
-        for (int required : requiredUniform)
-            maxUniformRandoms = max(maxUniformRandoms, required);
-        uniformRandoms.initialize<mm_float4>(cc, maxUniformRandoms, "uniformRandoms");
-        randomSeed.initialize<mm_int4>(cc, cc.getNumThreadBlocks()*64, "randomSeed");
-        vector<mm_int4> seed(randomSeed.getSize());
-        int rseed = integrator.getRandomNumberSeed();
-        // A random seed of 0 means use a unique one
-        if (rseed == 0)
-            rseed = osrngseed();
-        unsigned int r = (unsigned int) (rseed+1);
-        for (auto& s : seed) {
-            s.x = r = (1664525*r + 1013904223) & 0xFFFFFFFF;
-            s.y = r = (1664525*r + 1013904223) & 0xFFFFFFFF;
-            s.z = r = (1664525*r + 1013904223) & 0xFFFFFFFF;
-            s.w = r = (1664525*r + 1013904223) & 0xFFFFFFFF;
-        }
-        randomSeed.upload(seed);
-        ComputeProgram randomProgram = cc.compileProgram(CommonKernelSources::customIntegrator, defines);
-        randomKernel = randomProgram->createKernel("generateRandomNumbers");
-        randomKernel->addArg(maxUniformRandoms);
-        randomKernel->addArg(uniformRandoms);
-        randomKernel->addArg(randomSeed);
-        
-        // Create the kernel for computing kinetic energy.
-
-        stringstream computeKE;
-        for (int i = 0; i < perDofValues.size(); i++)
-            computeKE << tempType<<" perDof"<<cc.intToString(i)<<" = convertToTempType3(perDofValues"<<cc.intToString(i)<<"[index]);\n";
-        Lepton::ParsedExpression keExpression = Lepton::Parser::parse(integrator.getKineticEnergyExpression()).optimize();
-        computeKE << createPerDofComputation("", keExpression, integrator, "f", "", functionList, functionNames);
-        map<string, string> replacements;
-        replacements["COMPUTE_STEP"] = computeKE.str();
-        stringstream args;
-        for (int i = 0; i < perDofValues.size(); i++) {
-            string valueName = "perDofValues"+cc.intToString(i);
-            args << ", GLOBAL " << perDofType << "* RESTRICT " << valueName;
-        }
-        for (int i = 0; i < (int) tableTypes.size(); i++)
-            args << ", GLOBAL const " << tableTypes[i]<< "* RESTRICT table" << i;
-        replacements["PARAMETER_ARGUMENTS"] = args.str();
-        if (defines.find("LOAD_POS_AS_DELTA") != defines.end())
-            defines.erase("LOAD_POS_AS_DELTA");
-        ComputeProgram program = cc.compileProgram(cc.replaceStrings(CommonKernelSources::customIntegratorPerDof, replacements), defines);
-        kineticEnergyKernel = program->createKernel("computePerDof");
-        kineticEnergyKernel->addArg(cc.getPosq());
-        if (cc.getUseMixedPrecision())
-            kineticEnergyKernel->addArg(cc.getPosqCorrection());
-        else
-            kineticEnergyKernel->addArg(nullptr);
-        kineticEnergyKernel->addArg(integration.getPosDelta());
-        kineticEnergyKernel->addArg(cc.getVelm());
-        kineticEnergyKernel->addArg(cc.getLongForceBuffer());
-        kineticEnergyKernel->addArg(integration.getStepSize());
-        kineticEnergyKernel->addArg(globalValues);
-        kineticEnergyKernel->addArg(sumBuffer);
-        kineticEnergyKernel->addArg();
-        kineticEnergyKernel->addArg();
-        kineticEnergyKernel->addArg(uniformRandoms);
-        if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision())
-            kineticEnergyKernel->addArg(0.0);
-        else
-            kineticEnergyKernel->addArg(0.0f);
-        kineticEnergyKernel->addArg(perDofEnergyParamDerivs);
-        for (auto& array : perDofValues)
-            kineticEnergyKernel->addArg(array);
-        for (auto& array : tabulatedFunctions)
-            kineticEnergyKernel->addArg(array);
-
-        // Create a second kernel to sum the values.
-
-        program = cc.compileProgram(CommonKernelSources::customIntegrator, defines);
-        sumKineticEnergyKernel = program->createKernel(useDouble ? "computeDoubleSum" : "computeFloatSum");
-        sumKineticEnergyKernel->addArg(sumBuffer);
-        sumKineticEnergyKernel->addArg(summedValue);
-        sumKineticEnergyKernel->addArg(numAtoms);
-
-        // Delete the custom functions.
-
-        for (auto& function : functions)
-            delete function.second;
-    }
-
-    // Make sure all values (variables, parameters, etc.) are up to date.
-    
-    for (int i = 0; i < perDofValues.size(); i++) {
-        if (!deviceValuesAreCurrent[i]) {
-            if (useDouble)
-                perDofValues[i].upload(localPerDofValuesDouble[i]);
-            else
-                perDofValues[i].upload(localPerDofValuesFloat[i]);
-            deviceValuesAreCurrent[i] = true;
-        }
-        localValuesAreCurrent[i] = false;
-    }
-    double stepSize = integrator.getStepSize();
-    recordGlobalValue(stepSize, GlobalTarget(DT, dtVariableIndex), integrator);
-    for (int i = 0; i < (int) parameterNames.size(); i++) {
-        double value = context.getParameter(parameterNames[i]);
-        if (value != localGlobalValues[parameterVariableIndex[i]]) {
-            expressionSet.setVariable(parameterVariableIndex[i], value);
-            localGlobalValues[parameterVariableIndex[i]] = value;
-            deviceGlobalsAreCurrent = false;
-        }
-    }
-}
-
-ExpressionTreeNode CommonIntegrateCustomStepKernel::replaceDerivFunctions(const ExpressionTreeNode& node, ContextImpl& context) {
-    // This is called recursively to identify calls to the deriv() function inside global expressions,
-    // and replace them with a custom function that returns the correct value.
-    
-    const Operation& op = node.getOperation();
-    if (op.getId() == Operation::CUSTOM && op.getName() == "deriv") {
-        string param = node.getChildren()[1].getOperation().getName();
-        if (context.getParameters().find(param) == context.getParameters().end())
-            throw OpenMMException("The second argument to deriv() must be a context parameter");
-        needsEnergyParamDerivs = true;
-        return ExpressionTreeNode(new Operation::Custom("deriv", new DerivFunction(energyParamDerivs, param)));
-    }
-    else {
-        vector<ExpressionTreeNode> children;
-        for (auto& child : node.getChildren())
-            children.push_back(replaceDerivFunctions(child, context));
-        return ExpressionTreeNode(op.clone(), children);
-    }
-}
-
-void CommonIntegrateCustomStepKernel::findExpressionsForDerivs(const ExpressionTreeNode& node, vector<pair<ExpressionTreeNode, string> >& variableNodes) {
-    // This is called recursively to identify calls to the deriv() function inside per-DOF expressions,
-    // and record the code to replace them with.
-    
-    const Operation& op = node.getOperation();
-    if (op.getId() == Operation::CUSTOM && op.getName() == "deriv") {
-        string param = node.getChildren()[1].getOperation().getName();
-        int index;
-        for (index = 0; index < perDofEnergyParamDerivNames.size() && param != perDofEnergyParamDerivNames[index]; index++)
-            ;
-        if (index == perDofEnergyParamDerivNames.size())
-            perDofEnergyParamDerivNames.push_back(param);
-        string tempType = (cc.getSupportsDoublePrecision() ? "double3" : "float3");
-        variableNodes.push_back(make_pair(node, "make_"+tempType+"(energyParamDerivs["+cc.intToString(index)+"])"));
-        needsEnergyParamDerivs = true;
-    }
-    else {
-        for (auto& child : node.getChildren())
-            findExpressionsForDerivs(child, variableNodes);
-    }
-}
-
-void CommonIntegrateCustomStepKernel::execute(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid) {
-    ContextSelector selector(cc);
-    prepareForComputation(context, integrator, forcesAreValid);
-    IntegrationUtilities& integration = cc.getIntegrationUtilities();
-    int numAtoms = cc.getNumAtoms();
-    int numSteps = integrator.getNumComputations();
-    if (!forcesAreValid)
-        savedEnergy.clear();
-
-    // Loop over computation steps in the integrator and execute them.
-
-    for (int step = 0; step < numSteps; ) {
-        int nextStep = step+1;
-        int forceGroups = forceGroupFlags[step];
-        int lastForceGroups = context.getLastForceGroups();
-        bool haveForces = (!needsForces[step] || (forcesAreValid && lastForceGroups == forceGroups));
-        bool haveEnergy = (!needsEnergy[step] || savedEnergy.find(forceGroups) != savedEnergy.end());
-        if (!haveForces || !haveEnergy) {
-            if (forcesAreValid) {
-                if (savedForces.find(lastForceGroups) != savedForces.end() && validSavedForces.find(lastForceGroups) == validSavedForces.end()) {
-                    // The forces are still valid.  We just need a different force group right now.  Save the old
-                    // forces in case we need them again.
-
-                    cc.getLongForceBuffer().copyTo(savedForces[lastForceGroups]);
-                    validSavedForces.insert(lastForceGroups);
-                }
-            }
-            else
-                validSavedForces.clear();
-            
-            // Recompute forces and/or energy.  Figure out what is actually needed
-            // between now and the next time they get invalidated again.
-            
-            bool computeForce = (needsForces[step] || computeBothForceAndEnergy[step]);
-            bool computeEnergy = (needsEnergy[step] || computeBothForceAndEnergy[step]);
-            if (!computeEnergy && validSavedForces.find(forceGroups) != validSavedForces.end()) {
-                // We can just restore the forces we saved earlier.
-                
-                savedForces[forceGroups].copyTo(cc.getLongForceBuffer());
-                context.getLastForceGroups() = forceGroups;
-            }
-            else {
-                recordChangedParameters(context);
-                energy = context.calcForcesAndEnergy(computeForce, computeEnergy, forceGroups);
-                savedEnergy[forceGroups] = energy;
-                if (needsEnergyParamDerivs) {
-                    context.getEnergyParameterDerivatives(energyParamDerivs);
-                    if (perDofEnergyParamDerivNames.size() > 0) {
-                        for (int i = 0; i < perDofEnergyParamDerivNames.size(); i++)
-                            localPerDofEnergyParamDerivs[i] = energyParamDerivs[perDofEnergyParamDerivNames[i]];
-                        perDofEnergyParamDerivs.upload(localPerDofEnergyParamDerivs, true);
-                    }
-                }
-            }
-            forcesAreValid = true;
-        }
-        if (needsEnergy[step])
-            energy = savedEnergy[forceGroups];
-        if (needsGlobals[step] && !deviceGlobalsAreCurrent) {
-            // Upload the global values to the device.
-
-            globalValues.upload(localGlobalValues, true);
-            deviceGlobalsAreCurrent = true;
-        }
-        bool stepInvalidatesForces = invalidatesForces[step];
-        if (stepType[step] == CustomIntegrator::ComputePerDof && !merged[step]) {
-            kernels[step][0]->setArg(9, integration.prepareRandomNumbers(requiredGaussian[step]));
-            kernels[step][0]->setArg(8, integration.getRandom());
-            kernels[step][0]->setArg(10, uniformRandoms);
-            if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision())
-                kernels[step][0]->setArg(11, energy);
-            else
-                kernels[step][0]->setArg(11, (float) energy);
-            if (requiredUniform[step] > 0)
-                randomKernel->execute(numAtoms, 64);
-            kernels[step][0]->execute(numAtoms, 128);
-        }
-        else if (stepType[step] == CustomIntegrator::ComputeGlobal) {
-            expressionSet.setVariable(uniformVariableIndex, SimTKOpenMMUtilities::getUniformlyDistributedRandomNumber());
-            expressionSet.setVariable(gaussianVariableIndex, SimTKOpenMMUtilities::getNormallyDistributedRandomNumber());
-            expressionSet.setVariable(stepEnergyVariableIndex[step], energy);
-            recordGlobalValue(globalExpressions[step][0].evaluate(), stepTarget[step], integrator);
-        }
-        else if (stepType[step] == CustomIntegrator::ComputeSum) {
-            kernels[step][0]->setArg(9, integration.prepareRandomNumbers(requiredGaussian[step]));
-            kernels[step][0]->setArg(8, integration.getRandom());
-            kernels[step][0]->setArg(10, uniformRandoms);
-            if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision())
-                kernels[step][0]->setArg(11, energy);
-            else
-                kernels[step][0]->setArg(11, (float) energy);
-            if (requiredUniform[step] > 0)
-                randomKernel->execute(numAtoms, 64);
-            cc.clearBuffer(sumBuffer);
-            kernels[step][0]->execute(numAtoms, 128);
-            kernels[step][1]->execute(sumWorkGroupSize, sumWorkGroupSize);
-            if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision()) {
-                double value;
-                summedValue.download(&value);
-                recordGlobalValue(value, stepTarget[step], integrator);
-            }
-            else {
-                float value;
-                summedValue.download(&value);
-                recordGlobalValue(value, stepTarget[step], integrator);
-            }
-        }
-        else if (stepType[step] == CustomIntegrator::UpdateContextState) {
-            recordChangedParameters(context);
-            stepInvalidatesForces = context.updateContextState();
-        }
-        else if (stepType[step] == CustomIntegrator::ConstrainPositions) {
-            if (hasAnyConstraints) {
-                cc.getIntegrationUtilities().applyConstraints(integrator.getConstraintTolerance());
-                kernels[step][0]->execute(numAtoms);
-            }
-            cc.getIntegrationUtilities().computeVirtualSites();
-        }
-        else if (stepType[step] == CustomIntegrator::ConstrainVelocities) {
-            cc.getIntegrationUtilities().applyVelocityConstraints(integrator.getConstraintTolerance());
-        }
-        else if (stepType[step] == CustomIntegrator::IfBlockStart) {
-            if (!evaluateCondition(step))
-                nextStep = blockEnd[step]+1;
-        }
-        else if (stepType[step] == CustomIntegrator::WhileBlockStart) {
-            if (!evaluateCondition(step))
-                nextStep = blockEnd[step]+1;
-        }
-        else if (stepType[step] == CustomIntegrator::BlockEnd) {
-            if (blockEnd[step] != -1)
-                nextStep = blockEnd[step]; // Return to the start of a while block.
-        }
-        if (stepInvalidatesForces) {
-            forcesAreValid = false;
-            savedEnergy.clear();
-        }
-        step = nextStep;
-    }
-    recordChangedParameters(context);
-
-    // Update the time and step count.
-
-    cc.setTime(cc.getTime()+integrator.getStepSize());
-    cc.setStepCount(cc.getStepCount()+1);
-    cc.reorderAtoms();
-    if (cc.getAtomsWereReordered()) {
-        forcesAreValid = false;
-        validSavedForces.clear();
-    }
-    
-    // Reduce UI lag.
-
-    flushPeriodically(cc);
-}
-
-bool CommonIntegrateCustomStepKernel::evaluateCondition(int step) {
-    expressionSet.setVariable(uniformVariableIndex, SimTKOpenMMUtilities::getUniformlyDistributedRandomNumber());
-    expressionSet.setVariable(gaussianVariableIndex, SimTKOpenMMUtilities::getNormallyDistributedRandomNumber());
-    expressionSet.setVariable(stepEnergyVariableIndex[step], energy);
-    double lhs = globalExpressions[step][0].evaluate();
-    double rhs = globalExpressions[step][1].evaluate();
-    switch (comparisons[step]) {
-        case CustomIntegratorUtilities::EQUAL:
-            return (lhs == rhs);
-        case CustomIntegratorUtilities::LESS_THAN:
-            return (lhs < rhs);
-        case CustomIntegratorUtilities::GREATER_THAN:
-            return (lhs > rhs);
-        case CustomIntegratorUtilities::NOT_EQUAL:
-            return (lhs != rhs);
-        case CustomIntegratorUtilities::LESS_THAN_OR_EQUAL:
-            return (lhs <= rhs);
-        case CustomIntegratorUtilities::GREATER_THAN_OR_EQUAL:
-            return (lhs >= rhs);
-    }
-    throw OpenMMException("Invalid comparison operator");
-}
-
-double CommonIntegrateCustomStepKernel::computeKineticEnergy(ContextImpl& context, CustomIntegrator& integrator, bool& forcesAreValid) {
-    ContextSelector selector(cc);
-    prepareForComputation(context, integrator, forcesAreValid);
-    cc.clearBuffer(sumBuffer);
-    kineticEnergyKernel->setArg(8, cc.getIntegrationUtilities().getRandom());
-    kineticEnergyKernel->setArg(9, 0);
-    kineticEnergyKernel->execute(cc.getNumAtoms());
-    sumKineticEnergyKernel->execute(sumWorkGroupSize, sumWorkGroupSize);
-    if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision()) {
-        double ke;
-        summedValue.download(&ke);
-        return ke;
-    }
-    else {
-        float ke;
-        summedValue.download(&ke);
-        return ke;
-    }
-}
-
-void CommonIntegrateCustomStepKernel::recordGlobalValue(double value, GlobalTarget target, CustomIntegrator& integrator) {
-    switch (target.type) {
-        case DT:
-            if (value != localGlobalValues[dtVariableIndex])
-                deviceGlobalsAreCurrent = false;
-            expressionSet.setVariable(dtVariableIndex, value);
-            localGlobalValues[dtVariableIndex] = value;
-            cc.getIntegrationUtilities().setNextStepSize(value);
-            integrator.setStepSize(value);
-            break;
-        case VARIABLE:
-        case PARAMETER:
-            expressionSet.setVariable(target.variableIndex, value);
-            localGlobalValues[target.variableIndex] = value;
-            deviceGlobalsAreCurrent = false;
-            break;
-    }
-}
-
-void CommonIntegrateCustomStepKernel::recordChangedParameters(ContextImpl& context) {
-    if (!modifiesParameters)
-        return;
-    for (int i = 0; i < (int) parameterNames.size(); i++) {
-        double value = context.getParameter(parameterNames[i]);
-        if (value != localGlobalValues[parameterVariableIndex[i]])
-            context.setParameter(parameterNames[i], localGlobalValues[parameterVariableIndex[i]]);
-    }
-}
-
-void CommonIntegrateCustomStepKernel::getGlobalVariables(ContextImpl& context, vector<double>& values) const {
-    if (!globalValues.isInitialized()) {
-        // The data structures haven't been created yet, so just return the list of values that was given earlier.
-        
-        values = initialGlobalVariables;
-        return;
-    }
-    values.resize(numGlobalVariables);
-    for (int i = 0; i < numGlobalVariables; i++)
-        values[i] = localGlobalValues[globalVariableIndex[i]];
-}
-
-void CommonIntegrateCustomStepKernel::setGlobalVariables(ContextImpl& context, const vector<double>& values) {
-    if (numGlobalVariables == 0)
-        return;
-    if (!globalValues.isInitialized()) {
-        // The data structures haven't been created yet, so just store the list of values.
-        
-        initialGlobalVariables = values;
-        return;
-    }
-    for (int i = 0; i < numGlobalVariables; i++) {
-        localGlobalValues[globalVariableIndex[i]] = values[i];
-        expressionSet.setVariable(globalVariableIndex[i], values[i]);
-    }
-    deviceGlobalsAreCurrent = false;
-}
-
-void CommonIntegrateCustomStepKernel::getPerDofVariable(ContextImpl& context, int variable, vector<Vec3>& values) const {
-    ContextSelector selector(cc);
-    values.resize(perDofValues[variable].getSize());
-    const vector<int>& order = cc.getAtomIndex();
-    if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision()) {
-        if (!localValuesAreCurrent[variable]) {
-            perDofValues[variable].download(localPerDofValuesDouble[variable]);
-            localValuesAreCurrent[variable] = true;
-        }
-        for (int i = 0; i < (int) values.size(); i++) {
-            values[order[i]][0] = localPerDofValuesDouble[variable][i].x;
-            values[order[i]][1] = localPerDofValuesDouble[variable][i].y;
-            values[order[i]][2] = localPerDofValuesDouble[variable][i].z;
-        }
-    }
-    else {
-        if (!localValuesAreCurrent[variable]) {
-            perDofValues[variable].download(localPerDofValuesFloat[variable]);
-            localValuesAreCurrent[variable] = true;
-        }
-        for (int i = 0; i < (int) values.size(); i++) {
-            values[order[i]][0] = localPerDofValuesFloat[variable][i].x;
-            values[order[i]][1] = localPerDofValuesFloat[variable][i].y;
-            values[order[i]][2] = localPerDofValuesFloat[variable][i].z;
-        }
-    }
-}
-
-void CommonIntegrateCustomStepKernel::setPerDofVariable(ContextImpl& context, int variable, const vector<Vec3>& values) {
-    const vector<int>& order = cc.getAtomIndex();
-    localValuesAreCurrent[variable] = true;
-    deviceValuesAreCurrent[variable] = false;
-    if (cc.getUseDoublePrecision() || cc.getUseMixedPrecision()) {
-        localPerDofValuesDouble[variable].resize(values.size());
-        for (int i = 0; i < (int) values.size(); i++)
-            localPerDofValuesDouble[variable][i] = mm_double4(values[order[i]][0], values[order[i]][1], values[order[i]][2], 0);
-    }
-    else {
-        localPerDofValuesFloat[variable].resize(values.size());
-        for (int i = 0; i < (int) values.size(); i++)
-            localPerDofValuesFloat[variable][i] = mm_float4(values[order[i]][0], values[order[i]][1], values[order[i]][2], 0);
-    }
 }
 
 void CommonRemoveCMMotionKernel::initialize(const System& system, const CMMotionRemover& force) {
